Delete README.md
## Engineering Colleges case study
Education is fast becoming a very competitive sector with hundreds of institutions to choose from. It is a life-transforming experience for any student and it has to be a thoughtful decision. There are ranking agencies that do a survey of all the colleges to provide more insights to students and to identify different types of engineering colleges in the country to better understand the state of affairs.
## My aim is to.
- How many different types (clusters/segments) of colleges can be found from the data?
- How are these different types of colleges differ from each other? Please provide a profile summary of the clusters/segments you come up with. (These are your insights)
- Do you get slightly different solutions from two different techniques? How would you explain the difference?